requirements.txt
PyMySQL==0.9.3
python-dotenv
peewee
wheel